Contactless_Monitoring_System/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
PyJWT==2.8.0
bcrypt==4.1.3
qrcode[pil]==7.4.2
